requirements.txt
# Database Implementation Requirements
mysql-connector-python==8.2.0
pymysql==1.1.0
python-dotenv==1.0.0
tabulate==0.9.0
colorama==0.4.6
bcrypt==4.1.2